Updated server.go: New function findServer replaces actualStatus, further events are triggered, etc.

+++ pkg/cloud/resources/server/server.go
@@ -82,6 +82,12 @@ func (s *Service) Reconcile(ctx context.Context) (_ *ctrl.Result, err error) {
 		Image:             s.scope.HcloudMachine.Spec.ImageName,
 	})
 	if err != nil {
+		record.Warnf(s.scope.HcloudMachine,
+			"FailedEnsuringHcloudImage",
+			"Failed to ensure image for Hcloud server %s: %s",
+			s.scope.Name(),
+			err,
+		)
 		return nil, err
 	}
 
@@ -89,58 +95,49 @@ func (s *Service) Reconcile(ctx context.Context) (_ *ctrl.Result, err error) {
 		return &ctrl.Result{RequeueAfter: 2 * time.Second}, nil
 	}
 
-	if s.scope.IsBootstrapDataReady(s.scope.Ctx) {
+	if !s.scope.IsBootstrapDataReady(s.scope.Ctx) {
 		return &ctrl.Result{RequeueAfter: 15 * time.Second}, nil
 	}
 
 	// update current server
-	actualServers, err := s.actualStatus(ctx)
+	instance, err := s.findServer(ctx)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to refresh server status")
+		return nil, errors.Wrap(err, "failed to get server")
 	}
 
-	var actualServer *hcloud.Server
-
-	if len(actualServers) == 0 {
+	if instance == nil {
 		fmt.Printf("Wants to create server %s", s.scope.Machine.Name)
-		server, err := s.createServer(s.scope.Ctx, failureDomain, imageID)
+		instance, err = s.createServer(s.scope.Ctx, failureDomain, imageID)
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to create server")
 		}
-		if server != nil {
-			record.Eventf(
-				s.scope.HcloudMachine,
-				"SuccessfulCreate",
-				"Created new server with id %d",
-				server.ID,
-			)
-			actualServer = server
-		} else {
-			// In this case something was not ready yet
-			return &reconcile.Result{}, nil
-		}
-	} else if len(actualServers) == 1 {
-		actualServer = actualServers[0]
-	} else {
-		return nil, errors.New("found more than one actual servers")
+		record.Eventf(
+			s.scope.HcloudMachine,
+			"SuccessfulCreate",
+			"Created new server with id %d",
+			instance.ID,
+		)
 	}
-
-	if err := setStatusFromAPI(&s.scope.HcloudMachine.Status, actualServer); err != nil {
+	if err := setStatusFromAPI(&s.scope.HcloudMachine.Status, instance); err != nil {
 		return nil, errors.New("error setting status")
 	}
 
 	// wait for server being running
-	if actualServer.Status != hcloud.ServerStatusRunning {
-		s.scope.V(1).Info("server not in running state", "server", actualServer.Name, "status", actualServer.Status)
+	if instance.Status != hcloud.ServerStatusRunning {
+		s.scope.V(1).Info("server not in running state", "server", instance.Name, "status", instance.Status)
 		return &reconcile.Result{RequeueAfter: 2 * time.Second}, nil
 	}
 
-	providerID := fmt.Sprintf("hcloud://%d", actualServer.ID)
+	providerID := fmt.Sprintf("hcloud://%d", instance.ID)
 
 	if !s.scope.IsControlPlane() {
 		s.scope.HcloudMachine.Spec.ProviderID = &providerID
 		s.scope.HcloudMachine.Status.Ready = true
 		return nil, nil
+	}
+
+	if err := s.reconcileLoadBalancerAttachment(ctx, instance); err != nil {
+		return nil, errors.Wrap(err, "failed to add server to load balancer")
 	}
 
 	// check if at least one of the adresses is ready
@@ -156,10 +153,6 @@ func (s *Service) Reconcile(ctx context.Context) (_ *ctrl.Result, err error) {
 		})
 		if err != nil {
 			return nil, err
-		}
-
-		if err := s.addServerToLoadBalancer(ctx, actualServer); err != nil {
-			errors = append(errors, err)
 		}
 
 		if err := scope.IsControlPlaneReady(ctx, clientConfig); err != nil {
@@ -213,10 +206,11 @@ func (s *Service) createServer(ctx context.Context, failureDomain string, imageI
 
 	userDataInitial, err := s.scope.GetRawBootstrapData(ctx)
 	if err != nil {
-		if err == scope.ErrBootstrapDataNotReady {
-
-			return nil, nil
-		}
+		record.Warnf(
+			s.scope.HcloudMachine,
+			"FailedGetBootstrapData",
+			err.Error(),
+		)
 		return nil, fmt.Errorf("Failed to get raw bootstrap data: %s", err)
 	}
 
@@ -359,60 +353,64 @@ func (s *Service) createServer(ctx context.Context, failureDomain string, imageI
 	}
 	res, _, err := s.scope.HcloudClient().CreateServer(s.scope.Ctx, opts)
 	if err != nil {
+		record.Warnf(s.scope.HcloudMachine,
+			"FailedCreateHcloudServer",
+			"Failed to create Hcloud server %s: %s",
+			s.scope.Name(),
+			err,
+		)
 		return nil, fmt.Errorf("Error while creating Hcloud server %s: %s", &s.scope.HcloudMachine.Name, err)
 	}
+
 	return res.Server, nil
 }
 
 func (s *Service) Delete(ctx context.Context) (_ *ctrl.Result, err error) {
 	// update current servers
-	actualServers, err := s.actualStatus(ctx)
+	server, err := s.findServer(ctx)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to refresh server status")
 	}
-
-	var actionWait []*hcloud.Server
-	var actionShutdown []*hcloud.Server
-	var actionDelete []*hcloud.Server
-
-	for _, server := range actualServers {
-		switch status := server.Status; status {
-		case hcloud.ServerStatusRunning:
-			actionShutdown = append(actionShutdown, server)
-		case hcloud.ServerStatusOff:
-			actionDelete = append(actionDelete, server)
-		default:
-			actionWait = append(actionWait, server)
-		}
+	if server == nil {
+		s.scope.V(2).Info("Unable to locate Hcloud instance by ID or tags")
+		record.Warnf(s.scope.HcloudMachine, "NoInstanceFound", "Unable to find matching Hcloud instance for %s", s.scope.Name())
+		return &reconcile.Result{}, nil
 	}
 
-	// shutdown servers
-	for _, server := range actionShutdown {
+	err = s.deleteServerOfLoadBalancer(ctx, server)
+	if err != nil {
+		return &reconcile.Result{}, errors.Errorf("Error while deleting attached server of loadbalancer: %s", err)
+	}
+
+	switch status := server.Status; status {
+
+	case hcloud.ServerStatusRunning:
+
 		if _, _, err := s.scope.HcloudClient().ShutdownServer(ctx, server); err != nil {
-			return nil, errors.Wrap(err, "failed to shutdown server")
+			return &reconcile.Result{}, errors.Wrap(err, "failed to shutdown server")
 		}
-		actionWait = append(actionWait, server)
-	}
+		return &ctrl.Result{RequeueAfter: 5 * time.Second}, nil
 
-	// delete servers that need delete
-	for _, server := range actionDelete {
-		if server.Labels["machine_type"] == "control_plane" {
-			s.deleteServerOfLoadBalancer(ctx, server)
-		}
+	case hcloud.ServerStatusOff:
 
 		if _, err := s.scope.HcloudClient().DeleteServer(ctx, server); err != nil {
-			return nil, errors.Wrap(err, "failed to delete server")
+			record.Warnf(s.scope.HcloudMachine, "FailedDeleteHcloudServer", "Failed to delete Hcloud server %s", s.scope.Name())
+			return &reconcile.Result{}, errors.Wrap(err, "failed to delete server")
 		}
-	}
 
-	var result *ctrl.Result
-	if len(actionWait) > 0 {
-		result = &ctrl.Result{
-			RequeueAfter: 5 * time.Second,
-		}
-	}
+	default:
 
-	return result, nil
+		//actionWait
+		return &ctrl.Result{RequeueAfter: 5 * time.Second}, nil
+
+	}
+	record.Eventf(
+		s.scope.HcloudMachine,
+		"HcloudServerDeleted",
+		"Hcloud server %s deleted",
+		s.scope.Name(),
+	)
+	return &ctrl.Result{}, nil
 }
 
 func setStatusFromAPI(status *infrav1.HcloudMachineStatus, server *hcloud.Server) error {
@@ -454,7 +452,7 @@ func setStatusFromAPI(status *infrav1.HcloudMachineStatus, server *hcloud.Server
 	return nil
 }
 
-func (s *Service) addServerToLoadBalancer(ctx context.Context, server *hcloud.Server) error {
+func (s *Service) reconcileLoadBalancerAttachment(ctx context.Context, server *hcloud.Server) error {
 
 	var hasPrivateIP bool
 	if len(server.PrivateNet) > 0 {
@@ -470,6 +468,17 @@ func (s *Service) addServerToLoadBalancer(ctx context.Context, server *hcloud.Se
 
 	// If load balancer has not been attached to a network, then it cannot add a server with private IP
 	if hasPrivateIP == true && len(lb.PrivateNet) == 0 {
+		return nil
+	}
+
+	var alreadyAttached bool
+	for _, target := range lb.Targets {
+		if target.Server.Server.ID == server.ID {
+			alreadyAttached = true
+		}
+	}
+	// if server is already attached then return nil
+	if alreadyAttached {
 		return nil
 	}
 
@@ -493,6 +502,18 @@ func (s *Service) deleteServerOfLoadBalancer(ctx context.Context, server *hcloud
 	if err != nil {
 		return err
 	}
+
+	var stillAttached bool
+	for _, target := range lb.Targets {
+		if target.Server.Server.ID == server.ID {
+			stillAttached = true
+		}
+	}
+	// if server is not attached then return nil
+	if !stillAttached {
+		return nil
+	}
+
 	_, _, err = s.scope.HcloudClient().DeleteTargetServerOfLoadBalancer(ctx, lb, server)
 	if err != nil {
 		s.scope.V(2).Info("Could not delete server as target of load balancer", "Server", server.ID, "Load Balancer", lb.ID)
@@ -507,14 +528,23 @@ func (s *Service) deleteServerOfLoadBalancer(ctx context.Context, server *hcloud
 	return nil
 }
 
-// actualStatus gathers all matching server instances, matched by tag
-func (s *Service) actualStatus(ctx context.Context) ([]*hcloud.Server, error) {
+func (s *Service) findServer(ctx context.Context) (*hcloud.Server, error) {
 	opts := hcloud.ServerListOpts{}
 	opts.LabelSelector = utils.LabelsToLabelSelector(s.createLabels())
-	servers, err := s.scope.HcloudClient().ListServers(s.scope.Ctx, opts)
+	servers, err := s.scope.HcloudClient().ListServers(ctx, opts)
 	if err != nil {
 		return nil, err
 	}
+	if len(servers) > 1 {
+		record.Warnf(s.scope.HcloudMachine,
+			"MultipleInstances",
+			"Found %v instances of name %s",
+			len(servers),
+			s.scope.Name())
+		return nil, fmt.Errorf("Found %v servers with name %s", len(servers), s.scope.Name())
+	} else if len(servers) == 0 {
+		return nil, nil
+	}
 
-	return servers, nil
+	return servers[0], nil
 }
